pyproject.toml
[project]
name = "system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
]
